automatically update the internal versioning of MSVA.pm from the Changelog.
This is a little crufty (it might trip up people the first time they
bump versions directly from within a git repo, and cause a spurious
commit or two), but it has a few advantages:

 0) it is a simple, small change to make right now, and

 1) it means we can leave all the files in place, so the test harness
    can work in-place.

 2) our "upstream tarball" can continue to be cleanly built.

Probably the better long-term way to fix this is an "install" target
in the Makefile that does this substitution as it transfers the file
into place.  That target would need to know where to place perl
modules, among other things.  (see MS # 2684)